feat: Create missing_bridge_tags.sql
feat:  Query to get all the overpasses/underpasses/bridges in a given state that aren't properly tagged as such, since some routing algorithms take that into account for truck clearance, this would be a quick way to highlight high priority tags to fix in a region of interest. It would also pick up roads that intersect but don't share a node, when they otherwise should, which would also be a routing algorithm problem.